refactor(notes): remove side effects from notes action creator
