+++ src/dashboards/actions/notes.ts
@@ -1,25 +1,4 @@
-// Libraries
-import {get} from 'lodash'
-
-// Actions
-import {createCellWithView} from 'src/cells/actions/thunks'
-import {updateView} from 'src/views/actions/thunks'
-
-// Utils
-import {createView} from 'src/views/helpers'
-import {getByID} from 'src/resources/selectors'
-
-// Types
-import {
-  GetState,
-  MarkdownViewProperties,
-  NoteEditorMode,
-  ResourceType,
-  Dashboard,
-  View,
-} from 'src/types'
 import {NoteEditorState} from 'src/dashboards/reducers/notes'
-import {Dispatch} from 'react'
 
 export type Action =
   | CloseNoteEditorAction
@@ -67,28 +46,6 @@ export const setNote = (note: string): SetNoteAction => ({
   payload: {note},
 })
 
-export const createNoteCell = (dashboardID: string) => (
-  dispatch: Dispatch<Action | ReturnType<typeof createCellWithView>>,
-  getState: GetState
-) => {
-  const dashboard = getByID<Dashboard>(
-    getState(),
-    ResourceType.Dashboards,
-    dashboardID
-  )
-
-  if (!dashboard) {
-    throw new Error(`could not find dashboard with id "${dashboardID}"`)
-  }
-
-  const {note} = getState().noteEditor
-  const view = createView<MarkdownViewProperties>('markdown')
-
-  view.properties.note = note
-
-  return dispatch(createCellWithView(dashboard.id, view))
-}
-
 export interface ResetNoteStateAction {
   type: 'RESET_NOTE_STATE'
 }
@@ -108,55 +65,3 @@ export const setNoteState = (
   type: 'SET_NOTE_STATE',
   payload: noteState,
 })
-
-export const loadNote = (id: string) => (
-  dispatch: Dispatch<Action>,
-  getState: GetState
-) => {
-  const state = getState()
-  const currentViewState = getByID<View>(state, ResourceType.Views, id)
-
-  if (!currentViewState) {
-    return
-  }
-
-  const view = currentViewState
-
-  const note: string = get(view, 'properties.note', '')
-  const showNoteWhenEmpty: boolean = get(
-    view,
-    'properties.showNoteWhenEmpty',
-    false
-  )
-
-  const initialState = {
-    viewID: view.id,
-    note,
-    showNoteWhenEmpty,
-    mode: NoteEditorMode.Editing,
-  }
-
-  dispatch(setNoteState(initialState))
-}
-
-export const updateViewNote = (id: string) => (
-  dispatch: Dispatch<Action | ReturnType<typeof updateView>>,
-  getState: GetState
-) => {
-  const state = getState()
-  const {note, showNoteWhenEmpty} = state.noteEditor
-  const view = getByID<View>(state, ResourceType.Views, id)
-
-  if (view.properties.type === 'check') {
-    throw new Error(
-      `view type "${view.properties.type}" does not support notes`
-    )
-  }
-
-  const updatedView = {
-    ...view,
-    properties: {...view.properties, note, showNoteWhenEmpty},
-  }
-
-  return dispatch(updateView(view.dashboardID, updatedView))
-}
